move BlockTransformer to transformer package, rename SodiumBlockTransformer
